gmail: Extract the newClient() function.
The goal of this patch set is to add an alternate way
to authenticate against Gmail using JWT (JSON Web Token).

The auth token currently implemented in the code
conflict with the alternative JWT we want to support.

Also the JWT will not be storable in this program cache.

As a consequence this patch extract the minimum of code
for which an alternative code path using JWT must be
written.

The newClient() function has been extracted above
the Gmail struct in order to make the diff clear
and readable.